Fix warnings

+++ modules/FiAKEr/iARefDistCompute.cpp
@@ -136,7 +136,7 @@ void getBestMatches(iAFiberData const& fiber,
 			{
 				auto it = refCurveInfo.find(refFiberID);
 				iAFiberData refFiber(refTable, refFiberID, mapping, (it != refCurveInfo.end()) ? it->second : std::vector<iAVec3f>());
-				similarities[refFiberID].index = refFiberID;
+				similarities[refFiberID].index = static_cast<quint32>(refFiberID);
 				double curDissimilarity = getDissimilarity(fiber, refFiber, measuresToCompute[d].first, diagonalLength, maxLength);
 				if (std::isnan(curDissimilarity))
 				{
